+++ lib/trios/constants.dart
@@ -4,7 +4,7 @@ import 'package:intl/intl.dart';
 import 'package:trios/utils/extensions.dart';
 
 class Constants {
-  static const version = "1.1.0-dev07";
+  static const version = "1.1.0-dev08";
 
   static const appName = "TriOS";
   static const appTitle = "$appName v$version";