+++ lib/vram_estimator/charts/bar_chart.dart
@@ -1,10 +1,15 @@
+import 'package:collection/collection.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:trios/trios/app_state.dart';
 import 'package:trios/utils/extensions.dart';
 import 'package:trios/vram_estimator/graphics_lib_config_provider.dart';
 import 'package:trios/vram_estimator/models/gpu_info.dart';
 import 'package:trios/vram_estimator/models/graphics_lib_config.dart';
+import 'package:trios/vram_estimator/vram_estimator_page.dart';
+import 'package:trios/widgets/mod_icon.dart';
+import 'package:trios/widgets/moving_tooltip.dart';
 
 import '../../../utils/util.dart';
 import '../models/vram_checker_models.dart';
@@ -29,6 +34,7 @@ class VramBarChartState extends ConsumerState<VramBarChart> {
     final maxVramUsed = _calculateMostVramUse();
     final theme = Theme.of(context);
     graphicsLibConfig = ref.watch(graphicsLibConfigProvider);
+    final realMods = ref.watch(AppState.mods);
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
@@ -50,40 +56,71 @@ class VramBarChartState extends ConsumerState<VramBarChart> {
                     itemBuilder: (context, index) {
                       if (index > mods.length - 1) return const SizedBox();
                       final mod = mods[index];
+                      final realMod = realMods.firstWhereOrNull(
+                        (vramMod) => vramMod.id == mod.info.modInfo.id,
+                      );
                       final percentOfMax =
                           mod.bytesUsingGraphicsLibConfig(graphicsLibConfig) /
                           maxVramUsed;
                       final width = layoutConstraints.maxWidth * percentOfMax;
 
-                      return Container(
+                      final iconFilePath =
+                          realMod
+                              ?.findFirstEnabledOrHighestVersion
+                              ?.iconFilePath;
+
+                      return MovingTooltipWidget.framed(
+                        tooltipWidget:
+                            VramEstimatorPage.buildVramTopFilesTableWidget(
+                              theme,
+                              mod,
+                              graphicsLibConfig,
+                            ),
                         child: Card(
-                          clipBehavior: Clip.hardEdge,
+                          clipBehavior: Clip.antiAlias,
                           child: Padding(
                             padding: const EdgeInsets.only(top: 8.0),
                             child: Column(
                               crossAxisAlignment: CrossAxisAlignment.start,
                               children: [
-                                Padding(
-                                  padding: const EdgeInsets.symmetric(
-                                    horizontal: 8.0,
-                                  ),
-                                  child: Text(
-                                    mod.info.formattedName,
-                                    style: theme.textTheme.labelLarge,
-                                  ),
-                                ),
-                                Padding(
-                                  padding: const EdgeInsets.symmetric(
-                                    horizontal: 8.0,
-                                  ),
-                                  child: Text(
-                                    mod
-                                        .bytesUsingGraphicsLibConfig(
-                                          graphicsLibConfig,
-                                        )
-                                        .bytesAsReadableMB(),
-                                    style: theme.textTheme.labelMedium,
-                                  ),
+                                Row(
+                                  children: [
+                                    if (iconFilePath != null)
+                                      Padding(
+                                        padding: const EdgeInsets.only(
+                                          left: 8.0,
+                                        ),
+                                        child: ModIcon(iconFilePath, size: 28),
+                                      ),
+                                    Column(
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: [
+                                        Padding(
+                                          padding: const EdgeInsets.symmetric(
+                                            horizontal: 8.0,
+                                          ),
+                                          child: Text(
+                                            mod.info.formattedName,
+                                            style: theme.textTheme.labelLarge,
+                                          ),
+                                        ),
+                                        Padding(
+                                          padding: const EdgeInsets.symmetric(
+                                            horizontal: 8.0,
+                                          ),
+                                          child: Text(
+                                            mod
+                                                .bytesUsingGraphicsLibConfig(
+                                                  graphicsLibConfig,
+                                                )
+                                                .bytesAsReadableMB(),
+                                            style: theme.textTheme.labelMedium,
+                                          ),
+                                        ),
+                                      ],
+                                    ),
+                                  ],
                                 ),
                                 Opacity(
                                   opacity: 0.6,