+++ windows/flutter/generated_plugins.cmake
@@ -12,7 +12,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   url_launcher_windows
   window_manager
   window_size
-  windows_single_instance
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST